# entry/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Entry Service
    description: HTTP service for a multiple players game
    version: ""
host: localhost:8050
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /character:
        post:
            tags:
                - EntryCharacterService
            summary: creatCharacter EntryCharacterService
            operationId: EntryCharacterService#creatCharacter
            parameters:
                - name: CreatCharacterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/EntryCharacterServiceCreatCharacterRequestBody'
                    required:
                        - name
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/EntryCharacterServiceCreatCharacterResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - health
                            - experience
                            - inventoryId
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
            schemes:
                - http
    /character/{id}:
        get:
            tags:
                - EntryCharacterService
            summary: getCharacter EntryCharacterService
            operationId: EntryCharacterService#getCharacter
            parameters:
                - name: id
                  in: path
                  description: UUId of the Character
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/EntryCharacterServiceGetCharacterResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - health
                            - experience
                            - inventoryId
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
            schemes:
                - http
        delete:
            tags:
                - EntryCharacterService
            summary: deleteCharacter EntryCharacterService
            operationId: EntryCharacterService#deleteCharacter
            parameters:
                - name: id
                  in: path
                  description: UUId of the Character
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
            schemes:
                - http
        patch:
            tags:
                - EntryCharacterService
            summary: updateCharacter EntryCharacterService
            operationId: EntryCharacterService#updateCharacter
            parameters:
                - name: id
                  in: path
                  description: UUId of the Character
                  required: true
                  type: string
                - name: UpdateCharacterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/EntryCharacterServiceUpdateCharacterRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
            schemes:
                - http
    /inventory/{id}:
        get:
            tags:
                - EntryInventoryService
            summary: getInventory EntryInventoryService
            operationId: EntryInventoryService#getInventory
            parameters:
                - name: id
                  in: path
                  description: UUId of the inventory
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/EntryInventoryServiceGetInventoryResponseBody'
                        required:
                            - id
                            - userId
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
            schemes:
                - http
        patch:
            tags:
                - EntryInventoryService
            summary: updateInventory EntryInventoryService
            operationId: EntryInventoryService#updateInventory
            parameters:
                - name: id
                  in: path
                  description: UUId of the Inventory
                  required: true
                  type: string
                - name: UpdateInventoryRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/EntryInventoryServiceUpdateInventoryRequestBody'
                    required:
                        - itemsId
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
                "404":
                    description: Not Found response.
                    schema:
                        type: string
            schemes:
                - http
...
definitions:
    EntryCharacterServiceCreatCharacterRequestBody:
        title: EntryCharacterServiceCreatCharacterRequestBody
        type: object
        properties:
            description:
                type: string
                description: Description of the Character
                example: Numquam nemo esse ea sit porro.
            name:
                type: string
                description: Name of the Character
                example: Ea veritatis.
        example:
            description: Quibusdam iusto nobis quam rem.
            name: Omnis accusantium dolores suscipit.
        required:
            - name
    EntryCharacterServiceCreatCharacterResponseBody:
        title: EntryCharacterServiceCreatCharacterResponseBody
        type: object
        properties:
            description:
                type: string
                description: Description of the character
                example: Eum quam sint non necessitatibus earum sed.
            experience:
                type: integer
                description: The amount of experience value of the character
                example: 2383481284568078650
                format: int64
            health:
                type: integer
                description: The amount of health value of the character
                example: 7423830986562215167
                format: int64
            id:
                type: string
                description: UUId of the character
                example: Qui modi qui debitis odit dolorum.
            inventoryId:
                type: string
                description: UUId of the inventory record
                example: Sint quasi nemo sunt natus dicta blanditiis.
            name:
                type: string
                description: Name of the character
                example: Temporibus accusantium totam accusamus.
        example:
            description: Aut non debitis aut reiciendis.
            experience: 3016547019325695351
            health: 8297019746801853513
            id: Est labore et placeat aut nobis.
            inventoryId: Est et veritatis itaque quisquam delectus voluptates.
            name: Voluptates porro suscipit.
        required:
            - id
            - name
            - description
            - health
            - experience
            - inventoryId
    EntryCharacterServiceGetCharacterResponseBody:
        title: EntryCharacterServiceGetCharacterResponseBody
        type: object
        properties:
            description:
                type: string
                description: Description of the character
                example: Excepturi fuga perferendis quas odio.
            experience:
                type: integer
                description: The amount of experience value of the character
                example: 7687424890364925477
                format: int64
            health:
                type: integer
                description: The amount of health value of the character
                example: 3953714528485285362
                format: int64
            id:
                type: string
                description: UUId of the character
                example: Qui reprehenderit natus.
            inventoryId:
                type: string
                description: UUId of the inventory record
                example: Quo aut excepturi non ducimus.
            name:
                type: string
                description: Name of the character
                example: Ea temporibus.
        example:
            description: Dolorum sint vero.
            experience: 3399838487155674043
            health: 2435322486920218454
            id: Molestiae in est doloremque.
            inventoryId: Cum cupiditate dolores modi laudantium provident.
            name: Harum et eum.
        required:
            - id
            - name
            - description
            - health
            - experience
            - inventoryId
    EntryCharacterServiceUpdateCharacterRequestBody:
        title: EntryCharacterServiceUpdateCharacterRequestBody
        type: object
        properties:
            description:
                type: string
                description: Description of the Character
                example: Blanditiis minima ut alias.
            experience:
                type: integer
                description: The amount of experience value of the Character
                example: 3375384238095499247
                format: int64
            health:
                type: integer
                description: The amount of health value of the Character
                example: 4841819224652801619
                format: int64
            name:
                type: string
                description: Name of the Character
                example: Aut a fugit ipsum ex.
        example:
            description: Labore qui et.
            experience: 7387014515383921807
            health: 6028282123356271838
            name: Exercitationem voluptas nihil perferendis tenetur.
    EntryInventoryServiceGetInventoryResponseBody:
        title: EntryInventoryServiceGetInventoryResponseBody
        type: object
        properties:
            id:
                type: string
                description: UUId of the inventory
                example: A iure neque ea.
            itemsId:
                type: array
                items:
                    type: string
                    example: Ea vero mollitia consectetur iste optio.
                description: Array of uuid of items
                example:
                    - Molestias aut.
                    - Expedita sit suscipit magni vitae.
            userId:
                type: string
                description: UUId of the character that this inventory belongs to
                example: Suscipit dolore.
        example:
            id: In eveniet.
            itemsId:
                - Consequuntur ad mollitia.
                - Numquam quam fugiat aperiam recusandae reiciendis.
                - Dolor eum a.
            userId: Magnam aut dolorum sit.
        required:
            - id
            - userId
    EntryInventoryServiceUpdateInventoryRequestBody:
        title: EntryInventoryServiceUpdateInventoryRequestBody
        type: object
        properties:
            itemsId:
                type: array
                items:
                    type: string
                    example: Ut quo.
                description: Array of uuid of items
                example:
                    - Eos eveniet rem iste sed dicta.
                    - Et nisi qui deleniti.
                    - Aut consequatur mollitia maiores possimus tempora animi.
                    - Voluptas voluptas nostrum ut iusto nihil.
        example:
            itemsId:
                - Rerum officiis rerum sed.
                - Sint dignissimos.
                - Non excepturi.
                - Aperiam at rem voluptatem corporis hic.
        required:
            - itemsId
...